Add warning regarding neomodel
I have tried to get in contact with the maintainers of neomodel via their issues, with no response. 

There are easy issues that are open for over half a year (https://github.com/neo4j-contrib/neomodel/issues/627) and pull requests that do not get merged.
Currently the project is not usable.